[chore] utils/uploadImage.ts > 이미지 업로드 함수 분리
